+++ process/payment.php
@@ -23,15 +23,18 @@ $total=$_POST['total'];
 $method = $_POST['method'];
 $status = $_POST['status'];
 
-$myInvoice = mysqli_query($conn, "Select invoice from tb_transaksi where customer='".$customer."' and id_meja=".$meja." and `status`='not paid'");
-
-if($myInvoice->num_rows>0)
+if($meja!="")
 {
-    while($row=$myInvoice->fetch_assoc())
+    $myInvoice = mysqli_query($conn, "Select invoice from tb_transaksi where customer='".$customer."' and id_meja=".$meja." and `status`='not paid'");
+
+    if($myInvoice->num_rows>0)
     {
-        $invoice=$row['invoice'];
+        while($row=$myInvoice->fetch_assoc())
+        {
+            $invoice=$row['invoice'];
+        }
+        mysqli_query($conn, "delete from tb_transaksi where invoice='".$invoice."'");
     }
-    mysqli_query($conn, "delete from tb_transaksi where invoice='".$invoice."'");
 }
 else
 {
@@ -50,7 +53,7 @@ for($i=0;$i<count($data);$i++)
         }
         else
         {
-            $value=$value."('".$invoice."', '".Date('Y-m-d')."', '".Date('Y-m-d H:i:s')."', '".$customer."',".$_SESSION["id_kasir"].", ".$data[$i]->id.", ".$meja.", ".$data[$i]->qty.", ".$data[$i]->price.", ".$total.", '', 'cash', '".$description."', 'paid', '1')";
+            $value=$value."('".$invoice."', '".Date('Y-m-d')."', '".Date('Y-m-d H:i:s')."', '".$customer."',".$_SESSION["id_kasir"].", ".$data[$i]->id.", '', ".$data[$i]->qty.", ".$data[$i]->price.", ".$total.", '', 'cash', '".$description."', 'paid', '1')";
         }
     }
     else
@@ -61,10 +64,11 @@ for($i=0;$i<count($data);$i++)
         }
         else
         {
-            $value=$value.",('".$invoice."', '".Date('Y-m-d')."', '".Date('Y-m-d H:i:s')."', '".$customer."',".$_SESSION["id_kasir"].", ".$data[$i]->id.", ".$meja.", ".$data[$i]->qty.", ".$data[$i]->price.", ".$total.", '', 'cash', '".$description."' ,'paid', '1')";
+            $value=$value.",('".$invoice."', '".Date('Y-m-d')."', '".Date('Y-m-d H:i:s')."', '".$customer."',".$_SESSION["id_kasir"].", ".$data[$i]->id.", '', ".$data[$i]->qty.", ".$data[$i]->price.", ".$total.", '', 'cash', '".$description."' ,'paid', '1')";
         }
     }
 }
+
 
 $sql = mysqli_query($conn, "INSERT INTO tb_transaksi(`invoice`, `date`, `date_insert`, `customer`, `id_employee`, `id_menu`, `id_meja`, `qty`, `price`, `total_price`, `rest_total`, `method`, `description`, `status`, `method_order`) values".$value."");
